+++ src/main/java/dev/oscarrojas/order_manager/db/OrderItemData.java
@@ -1,3 +1,3 @@
 package dev.oscarrojas.order_manager.db;
 
-public record OrderItemData(ProductData product, int quantity, double unitPrice) {}
+public record OrderItemData(ProductData product, int quantity, long unitPrice) {}
